requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.5
flake8==3.9.2
idna==2.10
Jinja2==3.0.1
kombu==5.1.0
MarkupSafe==2.0.1
mccabe==0.6.1
plaid-python==7.4.0
prompt-toolkit==3.0.19
pycodestyle==2.7.0
pyflakes==2.3.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
